+++ examples/miniapps/sanic/giphynavigator/__main__.py
@@ -1,0 +1,8 @@
+"""Main module."""
+
+from .application import create_app
+
+
+if __name__ == '__main__':
+    app = create_app()
+    app.run(host='0.0.0.0', port=8000, debug=True)