+++ docs/main/changelog.rst
@@ -10,6 +10,7 @@ follows `Semantic versioning`_
 4.0.0
 -----
 - Add ``wiring`` feature.
+- Add ``sanic`` example.
 - Update ``aiohttp`` example.
 - Update ``flask`` example.